requirements.txt
python-dotenv==1.0.1
watchdog==3.0.0
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0
